Fixed a problem with filesystem stored images not linking correctly in the shopping pages Fixed an issue with filesystem warnings being delayed Added currency display to 'item-unitprice' and item-total' tags Enhanced internationalization support to include dynamic JavaScript interfaces Added 'category' display tags to the shopp('product') tag group Enhanced asset management to better support bulk deletes for db and filesystem storage Fixed a DB schema problem causing shipping fees added in the product editor to round to the nearest whole dollar value Changed defaults for buttons to only be used if no button value is specified (rather than when the button value option is empty) Added locale based checkout buttons and currency codes for PayPal Express Fixed an issue with invalid line items in Authorize.Net transactions due to long product names Adjusted the tab order for the order amount shipping rate UI Updated the internationalization files Fixed several url problems caused by different permalink structures and HTTPS connections
git-svn-id: file:///Users/jond/svngit/svnrepos/shopp/trunk@263 846ab242-8549-0410-9a75-dfdbdae9473e

+++ shopp/core/model/schema.sql
@@ -40,7 +40,7 @@ CREATE TABLE shopp_price (
 	price float(20,2) NOT NULL default '0.00',
 	saleprice float(20,2) NOT NULL default '0.00',
 	weight int(10) NOT NULL default '0',
-	shipfee int(10) NOT NULL default '0',
+	shipfee float(20,2) NOT NULL default '0',
 	stock int(10) NOT NULL default '0',
 	inventory enum('off','on') NOT NULL,
 	sale enum('off','on') NOT NULL,
